requirements.txt
websocket-client
requests
rich
fake-useragent
PySocks

